Move MainClassFinder to api package
