+++ jib-core/src/main/java/com/google/cloud/tools/jib/api/JavaContainerBuilder.java
@@ -20,7 +20,6 @@ import com.google.cloud.tools.jib.ProjectInfo;
 import com.google.cloud.tools.jib.configuration.LayerConfiguration;
 import com.google.cloud.tools.jib.filesystem.DirectoryWalker;
 import com.google.cloud.tools.jib.frontend.JavaEntrypointConstructor;
-import com.google.cloud.tools.jib.frontend.MainClassFinder;
 import com.google.common.annotations.VisibleForTesting;
 import java.io.IOException;
 import java.nio.file.Files;
